+++ container-engine-lib/docs/changelog.md
@@ -11,6 +11,7 @@
 * Implemented the methods: `CreateUserService`, `GetUserServices`, `StopUserServices` and `DestroyUserServices` in `DockerKurtosisBackend`
 * Added `NetworkingSidecarContainerTypeLabelValue` label value constant
 * Implemented `GetUserServiceLogs` method in `DockerKurtosisBackend`
+* Implemented `WaitForUserServiceHttpEndpointAvailability` method in `DockerKurtosisBackend`
 
 ### Breaking Changes
 * Added `networksIPAdrresses` in `Container` struct to store the different container ip addresses established on its Docker networks